+++ test/browser/features/fixtures/packages/vue-router/vite.config.js
@@ -1,0 +1,27 @@
+import { fileURLToPath, URL } from 'node:url'
+import path from 'node:path'
+
+import { defineConfig } from 'vite'
+import vue from '@vitejs/plugin-vue'
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [
+    vue(),
+  ],
+  resolve: {
+    alias: {
+      '@': fileURLToPath(new URL('./src', import.meta.url))
+    }
+  },
+  build: {
+    lib: {
+      entry: path.resolve(__dirname, 'src/main.js'),
+      name: 'VueRouter',
+      fileName: () => `bundle.js`
+    },
+  },
+  define: {
+    'process.env': {}
+  }
+})